remove legacy harbor code
